requirements.txt
Django==1.10.4
django-bootstrap3==7.1.0
python-dateutil==2.6.0
requests==2.12.3
